refactoe interface, remove unused props
