+++ components/Radio/RadioGroup.tsx
@@ -1,42 +1,6 @@
-import React, {
-  SFC,
-  ReactNode,
-  CSSProperties,
-  Component,
-  cloneElement,
-} from "react"
+import React, { Component, cloneElement } from "react"
 import _ from "lodash"
-import { RadioProps } from "./Radio"
-
-export interface RadioChangeEventTarget extends RadioProps {
-  checked: boolean
-}
-
-type size = "small" | "medium"
-
-export interface RadioChangeEvent {
-  target: RadioChangeEventTarget
-  stopPropagation: () => void
-  preventDefault: () => void
-  nativeEvent: MouseEvent
-}
-
-interface RadioGroupProps {
-  /** Value of the radio group */
-  value?: any
-  /** Custom styles for each radio button */
-  radioStyle?: CSSProperties
-  /** Function to call when radio is changed */
-  onChange?: (e: RadioChangeEvent) => void
-  /** default value of the group */
-  defaultValue?: any
-  /** Size of the radio items */
-  size?: size
-  /** children nodes */
-  children?: ReactNode
-  /** Whether the group is disabled */
-  disabled?: boolean
-}
+import { RadioGroupProps, RadioProps } from "./Interface"
 
 const renderRadio = (props, value, setValue) => {
   const propsToPass = _.omit(props, ["children"])
@@ -46,6 +10,7 @@ const renderRadio = (props, value, setValue) => {
     return cloneElement(Radio, {
       groupValue: value,
       setValue,
+      style: props.style,
       key: index,
       ...propsToPass,
     })
@@ -60,13 +25,13 @@ export default class RadioGroup extends Component<RadioGroupProps, any> {
   }
 
   state = {
-    value: this.props.value || this.props.defaultValue,
+    value: this.props.defaultValue || null,
   }
 
   setValue = value => {
     this.setState({ value })
   }
   render() {
-    return <div>{renderRadio(this.props, this.state.value, this.setValue)}</div>
+    return <>{renderRadio(this.props, this.state.value, this.setValue)}</>
   }
 }
